refactor common code for Prod and Sum; add dunders
